refactor(demo): split ContentView into smaller files
ContentView had grown to ~390 lines with five private subviews and a
clipboard helper inlined. Split into per-responsibility files so the
top-level view is just the scaffolding.

- Result/ — DominantColorView, PaletteGrid (+ PaletteChip),
  SwatchesView (+ SwatchCard), TimingsView.
- PhotoPickerLabel.swift — empty-state card / change-photo capsule.
- Support/ClipboardCopy.swift — `copyToPasteboard(_:copied:)` helper
  used by all three tap-to-copy cells. Animation duration tightened
  to 0.8s.

ContentView is 179 lines now (was 391).
